// src/app/demo/api/proyecto.ts
export class Proyecto {
    id?: number;
    titulo?: string;
    descripcion?: string;
    poblacionObjetivo?: string
    justificacion?: string;
    presupuesto?: string;
    resultadosEsperados?: string;
    observaciones?: string;
    fechaCreacion?: string;
    fechaActualizacion?: string;
    estado?: string;
    idUsuario?: number;
}